agent-kit/backend/requirements.txt
fastapi==0.109.1
uvicorn==0.23.2
python-dotenv==1.0.0
llama-index==0.10.7
web3==6.15.1
py-solc-x==2.0.2

